+++ @commitlint/cli/fixtures/comment-char/commitlint.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+	rules: {
+		'subject-empty': [2, 'never']
+	},
+	parserPreset: {
+        parserOpts: {
+            commentChar: '$'
+        }
+    },
+};