+++ @commitlint/cz-commitlint/src/types.ts
@@ -1,0 +1,6 @@
+import {RuleConfigCondition, RuleConfigSeverity} from '@commitlint/types';
+
+export type Rule =
+	| Readonly<[RuleConfigSeverity.Disabled]>
+	| Readonly<[RuleConfigSeverity, RuleConfigCondition]>
+	| Readonly<[RuleConfigSeverity, RuleConfigCondition, unknown]>;