+++ @commitlint/load/fixtures/recursive-extends-ts/first-extended/index.ts
@@ -1,0 +1,7 @@
+import type {UserConfig} from '../types';
+module.exports = {
+	extends: ['./second-extended'],
+	rules: {
+		one: [1, 'never', 'one']
+	}
+} as UserConfig;